backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["timofejkazakov <t.kazakov@agora.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.5"
django-celery-beat = "^2.6.0"
sentry-sdk = {extras = ["django"], version = "^2.6.0"}
drf-yasg = "^1.21.7"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
eventlet = "^0.36.1"
dependency-injector = "^4.42.0"
django-auditlog = "^3.0.0"
requests = "^2.32.3"
boto3 = "^1.35.39"
django-file-form = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
